requirements.txt
Flask>=2.0.0
Flask-Session>=0.4.0
PyJWT>=2.0.0
requests>=2.27.1
